# docs/api_docs.yaml
components:
  schemas:
...
    ChangePassword:
      properties:
        new_password:
          anyOf:
          - type: string
          - type: 'null'
          description: The new password of the user
          title: New Password
        old_password:
          anyOf:
          - type: string
          - type: 'null'
          description: The old password of the user
          title: Old Password
        username:
          anyOf:
          - type: string
          - type: 'null'
          description: The username of the user
          title: Username
      title: ChangePassword
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    Login:
      properties:
        password:
          anyOf:
          - type: string
          - type: 'null'
          description: The password of the user
          title: Password
        username:
          anyOf:
          - type: string
          - type: 'null'
          description: The username of the user
          title: Username
      title: Login
      type: object
...
    PageResult:
      description: PageResult info
      properties:
        count:
          anyOf:
          - type: integer
          - type: 'null'
          description: The total count of items
          title: Count
        page_number:
          anyOf:
          - type: integer
          - type: 'null'
          description: The page number
          title: Page Number
        page_size:
          anyOf:
          - type: integer
          - type: 'null'
          description: The page size
          title: Page Size
      title: PageResult
      type: object
...
    Register:
      properties:
        confirm_password:
          description: The confirm password of the user
          minLength: 6
          title: Confirm Password
          type: string
        email:
          anyOf:
          - type: string
          - type: 'null'
          description: The email of the user
          title: Email
        password:
          description: The password of the user
          minLength: 6
          title: Password
          type: string
        username:
          description: The username of the user
          maxLength: 50
          minLength: 2
          title: Username
          type: string
      required:
      - username
      - password
      - confirm_password
      title: Register
      type: object
...
    UserList:
      description: 用户列表响应
      properties:
        page_result:
          anyOf:
          - $ref: '#/components/schemas/PageResult'
          - type: 'null'
          description: 分页信息
        total:
          description: 用户总数
          title: Total
          type: integer
        users:
          description: 用户列表
          items:
            $ref: '#/components/schemas/UserSummary'
          title: Users
          type: array
      required:
      - users
      - total
      title: UserList
      type: object
    UserModel:
      description: 完整的用户信息模型
      properties:
        date_joined:
          description: 加入时间
          format: date-time
          title: Date Joined
          type: string
        email:
          anyOf:
          - type: string
          - type: 'null'
          description: 邮箱
          title: Email
        gmt_created:
          description: 创建时间
          format: date-time
          title: Gmt Created
          type: string
        gmt_updated:
          description: 更新时间
          format: date-time
          title: Gmt Updated
          type: string
        id:
          description: 用户ID
          title: Id
          type: string
        is_active:
          description: 是否激活
          title: Is Active
          type: boolean
        is_staff:
          description: 是否为员工
          title: Is Staff
          type: boolean
        is_superuser:
          description: 是否为超级用户
          title: Is Superuser
          type: boolean
        is_verified:
          description: 是否已验证
          title: Is Verified
          type: boolean
        role:
          description: 用户角色
          title: Role
          type: string
        username:
          description: 用户名
          title: Username
          type: string
      required:
      - id
      - username
      - role
      - is_active
      - is_verified
      - is_staff
      - is_superuser
      - date_joined
      - gmt_created
      - gmt_updated
      title: UserModel
      type: object
    UserSummary:
      description: 用户简略信息
      properties:
        date_joined:
          description: 加入时间
          format: date-time
          title: Date Joined
          type: string
        email:
          anyOf:
          - type: string
          - type: 'null'
          description: 邮箱
          title: Email
        id:
          description: 用户ID
          title: Id
          type: string
        is_active:
          description: 是否激活
          title: Is Active
          type: boolean
        role:
          description: 用户角色
          title: Role
          type: string
        username:
          description: 用户名
          title: Username
          type: string
      required:
      - id
      - username
      - role
      - is_active
      - date_joined
      title: UserSummary
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    APIKeyCookie:
      in: cookie
      name: session
      type: apiKey
info:
  title: API Server
  version: 0.1.0
openapi: 3.1.0
paths:
  /api/v1/auth/login:
    post:
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Login
...
  /api/v1/auth/register:
    post:
      operationId: register_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Register'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Register
  /api/v1/auth/user:
    get:
      operationId: get_user_api_v1_auth_user_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserModel'
          description: Successful Response
      security:
      - APIKeyCookie: []
      summary: Get User
  /api/v1/auth/users:
    get:
      operationId: list_users_api_v1_auth_users_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
          description: Successful Response
      security:
      - APIKeyCookie: []
      summary: List Users
  /api/v1/auth/users/change-password:
    post:
      operationId: change_password_api_v1_auth_users_change_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePassword'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Change Password
  /api/v1/auth/users/{user_id}:
    delete:
      operationId: delete_user_api_v1_auth_users__user_id__delete
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyCookie: []
      summary: Delete User
...