# web/generated/types/PageResult.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * API Server
 * OpenAPI spec version: 0.1.0
 */
import type { PageResultCount } from "./PageResultCount";
import type { PageResultPageNumber } from "./PageResultPageNumber";
import type { PageResultPageSize } from "./PageResultPageSize";

/**
 * PageResult info
 */
export interface PageResult {
  /** The total count of items */
  count?: PageResultCount;
  /** The page number */
  page_number?: PageResultPageNumber;
  /** The page size */
  page_size?: PageResultPageSize;
}