# web/generated/types/UserList.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * API Server
 * OpenAPI spec version: 0.1.0
 */
import type { UserListPageResult } from "./UserListPageResult";
import type { UserSummary } from "./UserSummary";

/**
 * 用户列表响应
 */
export interface UserList {
  /** 分页信息 */
  page_result?: UserListPageResult;
  /** 用户总数 */
  total: number;
  /** 用户列表 */
  users: UserSummary[];
}