# web/generated/types/UserModel.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * API Server
 * OpenAPI spec version: 0.1.0
 */
import type { UserModelEmail } from "./UserModelEmail";

/**
 * 完整的用户信息模型
 */
export interface UserModel {
  /** 加入时间 */
  date_joined: string;
  /** 邮箱 */
  email?: UserModelEmail;
  /** 创建时间 */
  gmt_created: string;
  /** 更新时间 */
  gmt_updated: string;
  /** 用户ID */
  id: string;
  /** 是否激活 */
  is_active: boolean;
  /** 是否为员工 */
  is_staff: boolean;
  /** 是否为超级用户 */
  is_superuser: boolean;
  /** 是否已验证 */
  is_verified: boolean;
  /** 用户角色 */
  role: string;
  /** 用户名 */
  username: string;
}