# web/generated/types/UserSummary.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * API Server
 * OpenAPI spec version: 0.1.0
 */
import type { UserSummaryEmail } from "./UserSummaryEmail";

/**
 * 用户简略信息
 */
export interface UserSummary {
  /** 加入时间 */
  date_joined: string;
  /** 邮箱 */
  email?: UserSummaryEmail;
  /** 用户ID */
  id: string;
  /** 是否激活 */
  is_active: boolean;
  /** 用户角色 */
  role: string;
  /** 用户名 */
  username: string;
}